# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meken API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/recovery-code": {
      "post": {
        "tags": [
          "identity"
        ],
        "summary": "Create Recovery Code",
        "description": "Create or rotate the account credential without changing device sessions.",
        "operationId": "create_recovery_code_v1_auth_recovery_code_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodeResponse"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/auth/recover": {
      "post": {
        "tags": [
          "identity"
        ],
        "summary": "Recover Account",
        "description": "Open an independent device session for the existing user; never merge identities.",
        "operationId": "recover_account_v1_auth_recover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "invalid_recovery_code: unknown, rotated or deleted account credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code; no request inputs are echoed."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "APIError"
      },
      "APIErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "APIErrorResponse"
      },
...
      "RecoveryCodeResponse": {
        "properties": {
          "recovery_code": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "title": "Recovery Code",
            "description": "Reusable 256-bit account credential. Shown only upon creation/rotation; keep it private in a password manager. Rotation invalidates the previous code."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "recovery_code"
        ],
        "title": "RecoveryCodeResponse"
      },
      "RecoveryRequest": {
        "properties": {
          "recovery_code": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "format": "password",
            "title": "Recovery Code",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "recovery_code"
        ],
        "title": "RecoveryRequest"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user_id"
        ],
        "title": "TokenResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}